[refactor] FacilityDto 중복 정의 제거 및 공통 클래스로 분리
