+++ src/main/java/com/wayble/server/wayblezone/exception/WaybleZoneErrorCase.java
@@ -10,7 +10,8 @@ public enum WaybleZoneErrorCase implements ErrorCase {
 
     WAYBLE_ZONE_NOT_FOUND(404, 2001, "해당 웨이블존을 찾을 수 없습니다."),
     INVALID_CATEGORY(400, 2002, "유효하지 않은 category 파라미터입니다."),
-    WAYBLE_ZONE_FACILITY_NOT_FOUND(404, 2003, "해당 웨이블존에 대한 시설 정보가 존재하지 않습니다.");
+    WAYBLE_ZONE_FACILITY_NOT_FOUND(404, 2003, "해당 웨이블존에 대한 시설 정보가 존재하지 않습니다."),
+    FACILITY_NOT_FOUND(404, 2002, "해당 웨이블존에 대한 시설 정보가 존재하지 않습니다.");
 
     private final Integer httpStatusCode;
     private final Integer errorCode;
